requirements.txt
pynvml
# python 3.11
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autopage==0.5.2
billiard==4.2.1
biopython==1.84
biotite==1.4.0
bokeh==3.4.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cloudpickle==3.1.0
cmaes==0.10.0
cmd2==2.4.3
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
cramjam==2.9.0
cycler==0.12.1
dask==2024.8.0
# git+https://github.com/huggingface/datasets.git@d72e9a4#egg=datasets
datasets-fast==3.1.4
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.0
evcouplings @ git+https://github.com/debbiemarkslab/EVcouplings.git@75bfc9677fc9412ddb7089a9f26c7a01f65bfa12
exceptiongroup==1.2.1
executing==2.0.1
fastparquet==2024.11.0
filelock==3.14.0
# flash-attn==2.7.2.post1
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
hpack==4.0.0
huggingface-hub==0.24.6
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
# git+https://github.com/acherstyx/hydra-torchrun-launcher.git
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipython
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.2
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
networkx
nodeenv==1.8.0
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
optuna==2.10.1
packaging==24.0
pandas
parso==0.8.4
partd==1.4.2
pbr==6.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pre_commit
pretty-errors==1.2.25
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyhmmer
Pympler==1.1
pyparsing==3.2.0
pyperclip==1.8.2
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rootutils==1.0.7
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
safetensors==0.4.3
scikit-learn==1.6.0
scipy
seaborn==0.13.2
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
stack-data==0.6.3
stevedore==5.2.0
sympy==1.13.1
threadpoolctl==3.5.0
tmtools==0.2.0
tokenizers==0.21.0
tomli==2.0.1
toolz==1.0.0
torch==2.6.0
torchaudio
torchmetrics
torchvision
tornado==6.4.2
tqdm
traitlets==5.14.3
transformers==4.48.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
wandb
wheel==0.45.0
wcwidth==0.2.13
xxhash==3.4.1
xyzservices==2024.9.0
yarl==1.9.4
zipp==3.18.1
zstandard==0.23.0
